View,Model,Controller의 기능 분리
